refactors models from urls.py
